nav between pages works; user already logged in goes to home screen
