+++ package.json
@@ -14,6 +14,7 @@
     "@expo/vector-icons": "^10.2.1",
     "@react-navigation/bottom-tabs": "^5.9.2",
     "@react-navigation/native": "^5.7.6",
+    "@react-navigation/stack": "^5.12.2",
     "expo": "^39.0.3",
     "expo-font": "~8.3.0",
     "expo-status-bar": "~1.0.2",
